Add config file, use goroutines to handle messages
This is a pretty ugly bunch of changes. I would have liked to separate
it out, but it would be a bunch of work for the marginal benefit of
didactic clarity. Instead, I'll just get a checkpoint.

I switched to a config file at /etc/subscribe-responder.toml since I now
have two passwords I'd have to pass to the commandline, and that bugs
me. I'm properly detecting existing emails and subscribing to state
changes, but there's still a decent amount of JMAP work to do.

I'm going to start working on the Listmonk changes next.