Overhaul to using two subscription lists.
We couldn't figure out a good way to keep track of whether or not we had
sent a welcome email to subscribers. This is what we came up with: have
two subscription lists. A subscriber requests admission to the first
list and campaigns are sent to the second list. We then watch for new
subscriptions on the first list, send a welcome transactional email,
then move their subscription to the second list.

This change includes new configuration options for doing just that. It
also removes sending a transactional email immediately on getting a new
subscriber email and instead waits for that to happen after they get
moved from the first list to the second.